requirements.txt
pyrogram==2.0.106
aiohttp==3.9.1
aiosqlite==0.19.0
python-dotenv==1.0.0
